Login token setup and guards on user's posts done

+++ app/Http/Controllers/Api/V1/PostController.php
@@ -7,6 +7,7 @@ use App\Http\Requests\StorePostRequest;
 use App\Http\Resources\PostResource;
 use App\Models\Post;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 
 class PostController extends Controller
 {
@@ -15,7 +16,9 @@ class PostController extends Controller
      */
     public function index()
     {
-        return PostResource::collection(Post::with('author')->paginate()); // ->get()
+        $user = request()->user();
+        $posts = $user->posts()->paginate();
+        return PostResource::collection($posts); // ->get()
     }
 
     /**
@@ -26,7 +29,7 @@ class PostController extends Controller
         //$data = $request->all();
         //$data = $request->only('title', 'body');
         $data = $request->validated();
-        $data['author_id'] = 1;
+        $data['author_id'] = $request->user()->id;
 
         $post = Post::create($data);
         return 
@@ -52,6 +55,13 @@ class PostController extends Controller
 
         //$post = Post::findOrFail($id);
 
+        //$user = request()->user(); 
+        /*
+        if($user->id != $post->author_id) {
+            abort(403, 'Access Forbidden - You are not authorized');
+        } */
+
+        abort_if(Auth::id() != $post->author_id, 403, 'Access Forbidden - You are not authorized');
         return response()->json(new PostResource($post), 201);
     }
 
@@ -60,6 +70,8 @@ class PostController extends Controller
      */
     public function update(Request $request, Post $post) // string $id
     {
+        abort_if(Auth::id() != $post->author_id, 403, 'Access Forbidden - You are not authorized');
+
         $data = $request->validate([
             'title' => 'required|string|min:2',
             'body' => 'required|string|min:2'
@@ -75,6 +87,8 @@ class PostController extends Controller
      */
     public function destroy(Post $post) //string $id
     {
+        abort_if(Auth::id() != $post->author_id, 403, 'Access Forbidden - You are not authorized');
+        
         $post->delete();
         return response()->noContent();
     }
